[#34] Kiosk Mode
Added Kiosk Mode, I'm looking for a lot of feedback here, but the basic
functionality is working:

- Is there a better way to brand this than "Kiosk Mode" -- I couldn't
think of a better word.
- I put in a default delay of 7.5 seconds between each action (show
punchline, show alt punchlines, show next joke), but this configurable
with the query param "delay" -- what's a good way to put this in the UI?
- Should we minimize the UI at all to communicate that it's in kiosk
mode?
- Should we show Qtips automatically?  I considered this, but I'm not
sure if it's a good idea.