[#34] Revisions to Kiosk Mode.
Rebranded to "Auto-jokels", added an auto-jokel footer, and fullscreen
functionality (only works with kiosk mode for now, it's actually kind of
slick so we might want to use it elsewhere.)

One thing I can't seem to fix now: the auto-jokels footer seems to push
the normal footer to below the bottom of the page.  I can't seem to fix
this.